# docs/api/openapi/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "CivicPulse API",
    "description": "AI Development-Needs Intelligence Layer \u2014 Code for Communities 2.0 Track 1. Intake, Multilingual NLP, Clustering, Geospatial, Evidence, Gap Detection, and Policy Simulation. Contract: docs/api/API_CONTRACT.md",
    "version": "0.1.0"
  },
  "paths": {
...
    "/requests": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Submit a citizen request (text or voice) \u2014 FR-001/002",
        "description": "Intake endpoint: runs STT, language detection, classification, NER, geocoding and clustering.",
        "operationId": "create_request_requests_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RequestCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestCreateResponse"
                }
              }
            }
          },
          "501": {
            "description": "Not Implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/NotImplementedResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/requests/{request_id}/audio": {
      "post": {
        "tags": [
          "requests"
        ],
        "summary": "Upload/attach audio for transcription \u2014 FR-002",
        "description": "Accepts citizen audio recording, runs speech transcription and links to citizen request.",
        "operationId": "upload_audio_requests__request_id__audio_post",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/Body_upload_audio_requests__request_id__audio_post"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AudioUploadResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/requests/{request_id}": {
      "get": {
        "tags": [
          "requests"
        ],
        "summary": "Citizen checks own submission status (reference ID) \u2014 FR-005",
        "description": "Look up submission status by reference code or ID.",
        "operationId": "get_request_requests__request_id__get",
        "parameters": [
          {
            "name": "request_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Request Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RequestStatusResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/clusters": {
      "get": {
        "tags": [
          "clusters"
        ],
        "summary": "List clusters, filterable by sector/district/status \u2014 FR-071 (analyst+)",
        "description": "List clusters with multi-factor filtering.",
        "operationId": "list_clusters_clusters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "name": "issue_type",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Issue Type"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Status"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClusterListResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/geospatial/clusters": {
      "get": {
        "tags": [
          "geospatial"
        ],
        "summary": "GeoJSON for map rendering \u2014 FR-029 (analyst+)",
        "description": "Returns valid GeoJSON points for clusters. Unresolved locations have geometry: None (FR-021/033).",
        "operationId": "cluster_geojson_geospatial_clusters_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "sector",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Sector"
            }
          },
          {
            "name": "district",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "District"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GeoJSONFeatureCollection"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/clusters/{cluster_id}/outcome": {
      "get": {
        "tags": [
          "outcomes"
        ],
        "summary": "Outcome measurement view \u2014 FR-064\u2013067 (analyst+)",
        "description": "Returns baseline vs synthetic followup indicators with correlation disclaimer (FR-067).",
        "operationId": "get_outcome_clusters__cluster_id__outcome_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "cluster_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Cluster Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutcomeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "AudioUploadResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "audio_url": {
            "type": "string",
            "title": "Audio Url"
          },
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "audio_url",
          "status"
        ],
        "title": "AudioUploadResponse"
      },
...
      "Body_upload_audio_requests__request_id__audio_post": {
        "properties": {
          "audio": {
            "type": "string",
            "format": "binary",
            "title": "Audio"
          }
        },
        "type": "object",
        "required": [
          "audio"
        ],
        "title": "Body_upload_audio_requests__request_id__audio_post"
      },
...
      "ClusterListResponse": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ClusterSummary"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total"
        ],
        "title": "ClusterListResponse"
      },
      "ClusterSummary": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "issue_type": {
            "type": "string",
            "title": "Issue Type"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "independent_demand_count": {
            "type": "integer",
            "title": "Independent Demand Count"
          },
          "raw_message_count": {
            "type": "integer",
            "title": "Raw Message Count"
          },
          "district": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "District"
          }
        },
        "type": "object",
        "required": [
          "id",
          "issue_type",
          "status",
          "independent_demand_count",
          "raw_message_count"
        ],
        "title": "ClusterSummary"
      },
...
      "ErrorBody": {
        "properties": {
          "code": {
            "type": "string",
            "title": "Code"
          },
          "message": {
            "type": "string",
            "title": "Message"
          },
          "details": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Details"
          }
        },
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "title": "ErrorBody"
      },
...
      "GeoJSONFeature": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "Feature"
          },
          "geometry": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Geometry"
          },
          "properties": {
            "type": "object",
            "title": "Properties"
          }
        },
        "type": "object",
        "title": "GeoJSONFeature"
      },
      "GeoJSONFeatureCollection": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type",
            "default": "FeatureCollection"
          },
          "features": {
            "items": {
              "$ref": "#/components/schemas/GeoJSONFeature"
            },
            "type": "array",
            "title": "Features"
          }
        },
        "type": "object",
        "title": "GeoJSONFeatureCollection"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "NotImplementedResponse": {
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ErrorBody"
          }
        },
        "type": "object",
        "title": "NotImplementedResponse",
        "description": "Uniform scaffold response for every unimplemented business endpoint."
      },
      "OutcomeResponse": {
        "properties": {
          "cluster_id": {
            "type": "integer",
            "title": "Cluster Id"
          },
          "baseline": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Baseline"
          },
          "followup": {
            "anyOf": [
              {
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Followup"
          },
          "is_synthetic": {
            "type": "boolean",
            "title": "Is Synthetic",
            "default": true
          },
          "disclaimer": {
            "type": "string",
            "title": "Disclaimer",
            "default": "Observed change, not proven causal impact (FR-067)"
          }
        },
        "type": "object",
        "required": [
          "cluster_id"
        ],
        "title": "OutcomeResponse"
      },
...
      "RequestCreate": {
        "properties": {
          "channel": {
            "type": "string",
            "pattern": "^(text|voice)$",
            "title": "Channel"
          },
          "language_hint": {
            "anyOf": [
              {
                "type": "string",
                "pattern": "^(hi|mr|en)$"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language Hint"
          },
          "text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Text"
          },
          "location_text": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Location Text"
          },
          "audio_base64": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Audio Base64"
          },
          "consent_ack": {
            "type": "boolean",
            "title": "Consent Ack"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "consent_ack"
        ],
        "title": "RequestCreate"
      },
      "RequestCreateResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "reference_code": {
            "type": "string",
            "title": "Reference Code"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "status",
          "reference_code"
        ],
        "title": "RequestCreateResponse"
      },
      "RequestStatusResponse": {
        "properties": {
          "request_id": {
            "type": "string",
            "title": "Request Id"
          },
          "reference_code": {
            "type": "string",
            "title": "Reference Code"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "language": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Language"
          },
          "transcript": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Transcript"
          },
          "created_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "request_id",
          "reference_code",
          "status"
        ],
        "title": "RequestStatusResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}